# Added option for disabling MSI autodiscover feature in azure_keyvault_secret lookup plugin